perf(spinner): load animation date after first render
