+++ packages/components/src/components/bal-spinner/bal-spinner.tsx
@@ -1,14 +1,23 @@
-import { Component, h, Host, Prop, Element, Watch } from '@stencil/core'
-import Lottie from 'lottie-web/build/player/lottie_light_html'
+import { Component, h, Host, Prop, Element, Watch, ComponentInterface } from '@stencil/core'
 import type { AnimationItem } from 'lottie-web/build/player/lottie_light_html'
-import { SpinnerAnimationData } from './bal-spinner.animation'
-import { flatten } from 'lottie-colorify'
+import { rIC } from '../../utils/helpers'
+import { Loggable, Logger, LogInstance } from '../../utils/log'
+
+type AnimationFunction = (el: HTMLElement, color: string) => AnimationItem
 
 @Component({
   tag: 'bal-spinner',
 })
-export class Spinner {
-  animation!: AnimationItem
+export class Spinner implements ComponentInterface, Loggable {
+  private animationItem!: AnimationItem
+  private animationFunction?: AnimationFunction
+
+  log!: LogInstance
+
+  @Logger('bal-spinner')
+  createLogger(log: LogInstance) {
+    this.log = log
+  }
 
   @Element() el!: HTMLElement
 
@@ -41,6 +50,7 @@ export class Spinner {
   componentDidUpdate() {
     this.resetAnimation()
   }
+
   componentDidLoad() {
     this.resetAnimation()
   }
@@ -49,26 +59,41 @@ export class Spinner {
     this.destroyAnimation()
   }
 
-  getColor(): string {
+  private getColor(): string {
     return this.inverted || this.color === 'white' ? '#ffffff' : '#151f6d'
   }
 
-  resetAnimation() {
+  private async resetAnimation() {
     this.destroyAnimation()
     if (!this.deactivated) {
-      this.animation = Lottie.loadAnimation({
-        container: this.el,
-        renderer: 'svg',
-        loop: true,
-        autoplay: true,
-        animationData: flatten(this.getColor(), SpinnerAnimationData(this.getColor())),
-      })
+      await this.loadAnimation()
+
+      if (this.animationFunction) {
+        this.animationFunction(this.el, this.getColor())
+      }
     }
   }
 
-  destroyAnimation() {
-    if (this.animation && this.animation.destroy) {
-      this.animation.destroy()
+  private async loadAnimation(): Promise<void> {
+    return new Promise((resolve, reject) => {
+      if (this.animationFunction) {
+        return resolve()
+      } else {
+        rIC(async () => {
+          import('./bal-spinner.animation')
+            .then(module => {
+              this.animationFunction = module.animate
+              resolve()
+            })
+            .catch(reject)
+        })
+      }
+    })
+  }
+
+  private destroyAnimation() {
+    if (this.animationItem && this.animationItem.destroy) {
+      this.animationItem.destroy()
     }
   }
