+++ packages/components/src/components/bal-spinner/test/bal-spinner.cy.html
@@ -15,7 +15,7 @@
         <bal-spinner></bal-spinner>
       </section>
 
-      <h2 class="title is-size-2">Small</h2>
+      <!-- <h2 class="title is-size-2">Small</h2>
       <section>
         <bal-spinner small></bal-spinner>
       </section>
@@ -25,7 +25,7 @@
         <div class="has-background-blue p-4">
           <bal-spinner inverted></bal-spinner>
         </div>
-      </section>
+      </section> -->
     </div>
   </body>
 </html>
